fix: Removed SSL configuration and infer SSL from host URL, add chromadb to localhost script

+++ local-dev.sh
@@ -64,9 +64,9 @@ cleanup() {
 
     # Stop Docker containers
     docker stop qdrant_dev || true
+    docker stop chromadb_dev || true
     cd ../..
     docker compose -f opensearch/docker-compose.yaml down
-    docker stop chromadb_dev || true
 }
 
 for sig in INT QUIT HUP TERM; do
@@ -79,6 +79,7 @@ trap cleanup EXIT
 
 # Stop any running vector db containers
 docker stop qdrant_dev || true
+docker stop chromadb_dev || true
 docker compose -f opensearch/docker-compose.yaml down
 
 # Create the databases directory if it doesn't exist
@@ -93,7 +94,9 @@ elif [ "${VECTOR_DB_PROVIDER:-QDRANT}" = "OPENSEARCH" ]; then
   docker compose -f opensearch/docker-compose.yaml up --detach
 elif [ "${VECTOR_DB_PROVIDER:-QDRANT}" = "CHROMADB" ]; then
   echo "Using ChromaDB as the vector database provider..."
-  docker run --name chromadb_dev --rm -d -p 8000:8000 -v $(pwd)/databases/chromadb_storage:/data chromadb/chroma
+  if [ "${CHROMADB_HOST:-}" = "localhost" ]; then
+    docker run --name chromadb_dev --rm -d -p 8000:8000 -v $(pwd)/databases/chromadb_storage:/data chromadb/chroma
+  fi
 else
   echo "Unsupported VECTOR_DB_PROVIDER: ${VECTOR_DB_PROVIDER}. Supported values are QDRANT or OPENSEARCH."
   exit 1
